requirements.txt
#absl-py==1.3.0
aioice==0.9.0
aiortc==1.5.0
altair==4.2.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
autoflake==2.0.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.6
bokeh==3.0.3
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cleanipynb==0.4.2
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.5
decorator==4.4.2
deepface==0.0.75
dnspython==2.3.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
fire==0.5.0
Flask==2.2.2
flatbuffers==23.1.4
fonttools==4.38.0
gast==0.4.0
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
ifaddr==0.2.0
imageio==2.27.0
imageio-ffmpeg==0.4.8
importanize==0.7.0
importlib-metadata==6.1.0
ipykernel==6.20.1
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.3
jupytext==1.14.4
kaggle==1.5.12
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==15.0.6.1
livelossplot==0.5.5
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
moviepy==1.0.3
mtcnn==0.1.1
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
#opencv-python==4.7.0.68
opencv-python-headless>=4.5.5.62
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
parso==0.8.3
pathlib2==2.3.7.post1
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.13
platformdirs==2.6.2
proglog==0.1.10
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
pyee==9.0.4
pyflakes==3.0.1
Pygments==2.14.0
pylibsrtp==0.8.0
pymongo==4.3.3
Pympler==1.0.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
#pywin32==305
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rich==13.3.3
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.0
scikit-plot==0.3.7
#scipy==1.10.0
scipy
seaborn==0.12.2
semver==3.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.6.2
streamlit==1.20.0
streamlit-webrtc==0.45.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
#tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.21
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.13
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wordcloud==1.8.2.2
wrapt==1.14.1
xgboost==1.7.4
xyzservices==2022.9.0
yarg==0.1.9
zipp==3.15.0
